agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "mcp>=1.9.0",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.8",
    "xai-sdk>=1.4.1",
]
